backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
requests==2.29.0
pillow==9.5.0
python-dotenv==1.0.0
pydantic==1.10.7
python-jose==3.3.0
